finished connection with server

+++ app/src/main/java/com/github/zottaa/note/details/NoteDetailsViewModel.kt
@@ -6,10 +6,12 @@ import androidx.lifecycle.ViewModel
 import com.github.zottaa.core.ClearViewModels
 import com.github.zottaa.core.Now
 import com.github.zottaa.core.Screen
+import com.github.zottaa.core.SharedPreferencesRepository
 import com.github.zottaa.main.Navigation
 import com.github.zottaa.note.core.CategoryRepository
 import com.github.zottaa.note.core.NoteLiveDataWrapper
 import com.github.zottaa.note.core.NotesRepository
+import com.github.zottaa.note.core.RemoteNotesRepository
 import com.github.zottaa.note.list.CategoryUi
 import com.github.zottaa.note.list.ListLiveDataWrapper
 import com.github.zottaa.note.list.NoteUi
@@ -22,6 +24,8 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
 class NoteDetailsViewModel(
+    private val sharedPreferencesRepository: SharedPreferencesRepository,
+    private val remoteNotesRepository: RemoteNotesRepository.Edit,
     private val categoryRepository: CategoryRepository.All,
     private val noteLiveDataWrapper: NoteLiveDataWrapper.Mutable,
     private val noteListLiveDataWrapper: ListLiveDataWrapper.Edit,
@@ -50,10 +54,20 @@ class NoteDetailsViewModel(
 
     fun deleteNote(noteId: Long) {
         viewModelScope.launch(dispatcher) {
-            repository.deleteNote(noteId)
-            withContext(dispatcherMain) {
-                noteListLiveDataWrapper.delete(noteId)
-                comeback()
+            val categoryId = repository.note(noteId).categoryId()
+            val result = remoteNotesRepository.delete(
+                sharedPreferencesRepository.getUserId(),
+                noteId,
+                categoryId
+            )
+            if (result.isSuccess()) {
+                repository.deleteNote(noteId)
+                withContext(dispatcherMain) {
+                    noteListLiveDataWrapper.delete(noteId)
+                    comeback()
+                }
+            } else {
+                println(result.message())
             }
         }
     }
@@ -66,18 +80,30 @@ class NoteDetailsViewModel(
         currentCategory: Long
     ) {
         viewModelScope.launch(dispatcher) {
-            repository.updateNote(noteId, newTitle, newText, categoryId)
-            withContext(dispatcherMain) {
-                if (categoryId != currentCategory && currentCategory != ALL_CATEGORY)
-                    noteListLiveDataWrapper.delete(noteId)
-                else
-                    noteListLiveDataWrapper.updateOrCreate(
-                        noteId,
-                        newTitle,
-                        newText,
-                        now.timeInMillis(),
-                        categoryId
-                    )
+            val result = remoteNotesRepository.update(
+                sharedPreferencesRepository.getUserId(),
+                noteId,
+                newTitle,
+                newText,
+                now.timeInMillis(),
+                categoryId
+            )
+            if (result.isSuccess()) {
+                repository.updateNote(noteId, newTitle, newText, categoryId)
+                withContext(dispatcherMain) {
+                    if (categoryId != currentCategory && currentCategory != ALL_CATEGORY)
+                        noteListLiveDataWrapper.delete(noteId)
+                    else
+                        noteListLiveDataWrapper.updateOrCreate(
+                            noteId,
+                            newTitle,
+                            newText,
+                            now.timeInMillis(),
+                            categoryId
+                        )
+                }
+            } else {
+                println(result.message())
             }
         }
     }
